Merge con metodos de news y pom actualizado

+++ Polo/src/main/java/com/Polo/controller/SuscriptionController.java
@@ -1,5 +1,6 @@
 package com.Polo.controller;
 
+import java.util.Map;
 import java.util.Optional;
 
 import org.springframework.http.HttpStatus;
@@ -15,8 +16,9 @@ import org.springframework.web.bind.annotation.RestController;
 
 import com.Polo.model.Suscription;
 import com.Polo.service.SuscriptionService;
-import com.Polo.service.UserService;
+import com.Polo.userDataSession.SessionUtils;
 
+import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
 
 @RestController
@@ -25,7 +27,6 @@ import lombok.RequiredArgsConstructor;
 public class SuscriptionController {
 
     private final SuscriptionService suscriptionService;
-    private final UserService userService;
 
     // crear suscriptor
     @PostMapping("/create")
@@ -65,22 +66,22 @@ public class SuscriptionController {
 
     // elimiar suscriptor por admin
     @DeleteMapping("/deleteSuscriptor/{adminName}")
-    public ResponseEntity<String> deleteSuscriptorByAdmin(@PathVariable String adminName, @RequestParam String subEmail) {
+    public ResponseEntity<String> deleteSuscriptorByAdmin(@PathVariable String adminName, @RequestParam String subEmail, HttpSession session) {
 
-        // Validar si el usuario que está solicitando es ADMIN
-        if (!userService.isAdmin(adminName)) {
-            return new ResponseEntity<>("User Admin isn't ADMIN", HttpStatus.FORBIDDEN);
+        Map<String, Object> sessionData = SessionUtils.getUserSession(session);
+        String role = sessionData.get("role").toString();
+
+        if ("ADMIN".equals(role)) {
+            // Intentar eliminar al usuario
+            boolean check = suscriptionService.deleteSuscriptorByMail(subEmail);
+
+            if (check) {
+                return new ResponseEntity<>("Suscriptior deleted successfully", HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>("Suscriptor cannot be deleted", HttpStatus.NOT_FOUND);
+            }
         } else {
-            System.out.println("El admin name esta correcto");
-        }
-
-        // Intentar eliminar al usuario
-        boolean check = suscriptionService.deleteSuscriptorByMail(subEmail);
-
-        if (check) {
-            return new ResponseEntity<>("Suscriptior deleted successfully", HttpStatus.OK);
-        } else {
-            return new ResponseEntity<>("Suscriptor cannot be deleted", HttpStatus.NOT_FOUND);
+            return new ResponseEntity<>("Usuario sin permiso para esta accion", HttpStatus.NOT_FOUND);
         }
     }
